requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
google.cloud
pdfplumber
vertexai